I added a text_length block.  It has a bug (which I think is in the parser) - if a string is surrounded by double quotes, the length is two less than it should be.
I then decided that I needed to type the variables (now that we have
number, boolean, and string variables).  It is mostly done and working
now.  I now need to make changing a variable name a real change to pick up
those too - right now the variable can change names without triggering a
type change (which it needs if it used to be an instance of a variable
with a different type, and now is either a new variable or linked to a
variable with a different type).  But I want to check this version in
first.