took out all autosaving by Blockscad, replaced with a system that better detects when a save is needed, and if the user has not explicitly saved, throws up a dialog box asking "do you want to save?".
Fixed the firefox xmlns namespace bug that was corrupting projects (ouch).

Fixed David Jin's reported bug that caused blocks to draw poorly on load
if there was an open text_editor on a number block in a cylinder already
in the work area before load.

Improved the internal "which project is this" tracking so that we should
get fewer duplicates of projects when saving (it was related to page
refresh, I think).

Added bootbox for confirm/delete modals, and lie to polyfill promises for
older browsers.